requirements.txt
numpy==1.19.5
Werkzeug==1.0.1
Flask==1.1.2
tensorflow==2.4.1
Pillow==8.1.2
gunicorn>=20.0.0
